+++ alv-portal-ui/src/app/candidate/candidate-search-profiles/candidate-search-profiles/candidate-search-profiles.component.html
@@ -26,8 +26,7 @@
       baseRouterLink="/candidate-search"
       (deleted)="onDeleteProfile($event)">
     </alv-search-profile-item>
-
-
+    
   </ng-container>
   <ng-container *ngIf="candidateSearchProfiles && !candidateSearchProfiles.length">
     <div class="no-profiles p-3">
