+++ alv-portal-ui/src/app/candidate/candidate-search-profiles/candidate-search-profiles/candidate-search-profiles.component.ts
@@ -35,7 +35,6 @@ export class CandidateSearchProfilesComponent implements OnInit {
     this.onScroll();
   }
 
-
   onScroll() {
     this.authenticationService.getCurrentUser().pipe(
       take(1),
