+++ alv-portal-ui/src/app/shared/search-profiles/modal-config.types.ts
@@ -11,9 +11,9 @@ export function getJobAdDeleteConfirmModalConfig(profileName: string): ConfirmMo
 
 export function getCandidateDeleteConfirmModalConfig(profileName: string): ConfirmModalConfig {
   return {
-    title: 'portal.job-ad-search-profiles.delete-confirmation-modal.title',
-    content: 'portal.job-ad-search-profiles.delete-confirmation-modal.question',
-    confirmLabel: 'portal.job-ad-search-profiles.delete-confirmation-modal.button.confirm',
+    title: 'portal.candidate-search-profiles.delete-confirmation-modal.title',
+    content: 'portal.candidate-search-profiles.delete-confirmation-modal.question',
+    confirmLabel: 'portal.candidate-search-profiles.delete-confirmation-modal.button.confirm',
     contentParams: { profileName: profileName }
   };
 }
